More ConflictSet and Var to separate modules
In preparation for making ConflictSet abstract.
